+++ client/src/ui/components/worldmap/armies/ArmyMenu.tsx
@@ -22,9 +22,9 @@ const EXPLORATION_REWARD_RESOURCE_AMOUNT: number = 20;
 const EXPLORATION_COSTS: Resource[] = [
   {
     resourceId: 254,
-    amount: 300000,
+    amount: 30000,
   },
-  { resourceId: 255, amount: 150000 },
+  { resourceId: 255, amount: 15000 },
 ];
 
 export const ArmyMenu = ({ entityId }: ArmyMenuProps) => {
@@ -128,7 +128,7 @@ export const ArmyMenu = ({ entityId }: ArmyMenuProps) => {
   const { getResourcesFromBalance } = useResources();
 
   const explorationCosts = useMemo(() => {
-    const foodBalance = entityOwner ? getResourcesFromBalance(entityOwner.entity_owner_id).resources : [];
+    const foodBalance = entityOwner ? getResourcesFromBalance(entityOwner.entity_owner_id) : [];
     return EXPLORATION_COSTS.map((res) => {
       return {
         ...res,
