+++ sdk/packages/eternum/src/types/common.ts
@@ -11,6 +11,12 @@ export enum DESTINATION_TYPE {
   BANK,
 }
 
+export enum ENTITY_TYPE {
+  DONKEY,
+  TROOP,
+  UNKNOWN,
+}
+
 export interface CombatResultInterface {
   attackerRealmEntityId: bigint;
   targetRealmEntityId: bigint;
@@ -93,7 +99,7 @@ export interface RoadInterface {
   usageLeft: number;
 }
 
-export interface CaravanInterface {
+export interface EntityInterface {
   entityId: bigint;
   blocked: boolean | undefined;
   arrivalTime: number | undefined;
@@ -105,6 +111,7 @@ export interface CaravanInterface {
   position: Position | undefined;
   homePosition: Position | undefined;
   resources: Resource[];
+  entityType: ENTITY_TYPE;
 }
 
 /// REALMS
